backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Google AI
google-generativeai>=0.3.0

# Data processing
pandas>=2.1.0

# Retry logic
tenacity>=8.2.0

# CLI/logging (optional)
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
